refactor: both value precede variants to a single file
