+++ crates/huub/src/constraints.rs
@@ -12,8 +12,7 @@ pub mod int_linear;
 pub mod int_pow;
 pub mod int_table;
 pub mod int_times;
-pub mod seq_precede_chain;
-pub mod value_precede_chain;
+pub mod int_value_precede;
 
 use std::{
 	error::Error,
