+++ crates/huub/src/lib.rs
@@ -58,8 +58,7 @@ use crate::{
 		int_pow::IntPow,
 		int_table::IntTable,
 		int_times::IntTimes,
-		seq_precede_chain::SeqPrecedeChain,
-		value_precede_chain::ValuePrecedeChain,
+		int_value_precede::{SeqPrecedeChain, ValuePrecedeChain},
 		BoxedConstraint, Constraint, SimplificationStatus,
 	},
 	flatzinc::{FlatZincError, FlatZincStatistics, FznModelBuilder},
