+++ crates/huub/src/reformulate.rs
@@ -31,8 +31,7 @@ use crate::{
 		int_pow::IntPow,
 		int_table::IntTable,
 		int_times::IntTimes,
-		seq_precede_chain::SeqPrecedeChain,
-		value_precede_chain::ValuePrecedeChain,
+		int_value_precede::{SeqPrecedeChain, ValuePrecedeChain},
 		BoxedConstraint, BoxedPropagator, Constraint, SimplificationStatus,
 	},
 	helpers::linear_transform::LinearTransform,
